+++ android/YaSeArmo/app/src/main/java/com/masavi/yasearmo/MainMenuActivity.java
@@ -1,78 +1,26 @@
 package com.masavi.yasearmo;
 
 import android.content.Intent;
-import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
 
-import com.google.android.gms.auth.api.Auth;
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
-import com.google.android.gms.common.ConnectionResult;
-import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.common.api.ResultCallback;
-import com.google.android.gms.common.api.Status;
-import com.google.firebase.auth.FirebaseAuth;
+import butterknife.ButterKnife;
+import butterknife.OnClick;
 
 public class MainMenuActivity extends AppCompatActivity {
-
-    private GoogleApiClient mGoogleApiClient;
-    private Button mLogOutBtn;
-
-    FirebaseAuth mAuth;
-    private FirebaseAuth.AuthStateListener mAuthListener;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main_menu);
-
-        mAuth = FirebaseAuth.getInstance();
-        mAuthListener = new FirebaseAuth.AuthStateListener() {
-            @Override
-            public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
-
-                if (firebaseAuth.getCurrentUser() == null){
-                    startActivity(new Intent(MainMenuActivity.this, MainActivity.class));
-                }
-            }
-        };
-
-        // Configure Google Sign In
-        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.default_web_client_id))
-                .requestEmail()
-                .build();
-
-        mGoogleApiClient = new GoogleApiClient.Builder(getApplicationContext())
-                .enableAutoManage(this, new GoogleApiClient.OnConnectionFailedListener() {
-                    @Override
-                    public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
-                        Toast.makeText(MainMenuActivity.this, "Hay un error!", Toast.LENGTH_SHORT).show();
-                    }
-                })
-                .addApi(Auth.GOOGLE_SIGN_IN_API, gso)
-                .build();
-
-        mLogOutBtn = (Button) findViewById(R.id.btn_logout);
-
-        mLogOutBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // Firebase Sign Out
-                mAuth.signOut();
-
-                // Google sign out
-                Auth.GoogleSignInApi.signOut(mGoogleApiClient);
-            }
-        });
+        ButterKnife.bind(this);
     }
 
-    @Override
-    protected void onStart() {
-        super.onStart();
-        mAuth.addAuthStateListener(mAuthListener);
+    @OnClick(R.id.btn_perfil)
+    public void onClick(View view) {
+        startActivity(new Intent(MainMenuActivity.this, ProfileActivity.class));
     }
+
+
 }
