+++ android/YaSeArmo/app/src/main/java/com/masavi/yasearmo/util/Constants.java
@@ -1,0 +1,10 @@
+package com.masavi.yasearmo.util;
+
+/**
+ * Created by Masavi on 24/5/2017.
+ */
+
+public class Constants {
+    public static final int RC_SIGN_IN = 9001;
+    public static final String TAG = "MAIN_ACTIVITY";
+}
