+++ android/YaSeArmo/app/src/main/res/layout/activity_login.xml
@@ -1,18 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.masavi.yasearmo.MainActivity">
+    android:orientation="vertical"
+    tools:context="com.masavi.yasearmo.LoginActivity">
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/btn_google"
-        android:layout_width="368dp"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        tools:layout_editor_absoluteY="0dp"
-        tools:layout_editor_absoluteX="8dp">
+        android:layout_margin="50dp"
+        android:layout_gravity="center"
+        >
 
     </com.google.android.gms.common.SignInButton>
 
-</android.support.constraint.ConstraintLayout>
+</LinearLayout>
