+++ android/YaSeArmo/app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
+    <color name="colorPrimary">#5d006a</color>
+    <color name="colorPrimaryDark">#3e0047</color>
     <color name="colorAccent">#FF4081</color>
 </resources>
